backend/requirements.txt
fastapi
uvicorn
pydantic[email]
sentence-transformers
scikit-learn
nltk
motor
passlib[bcrypt]
bcrypt
python-jose[cryptography]
python-multipart
python-dotenv
email-validator
certifi
numpy
dnspython
yake
requests
